Result of `go test -bench=. -cpuprofile=cpu.out`

**Before**
```
goos: linux
goarch: amd64
pkg: github.com/andrepputra/benchmarking
BenchmarkSimulateLootRNG-12          144           8226728 ns/op
```

**After**
```
goos: linux
goarch: amd64
pkg: github.com/andrepputra/benchmarking
BenchmarkSimulateLootRNG-12         4651            275755 ns/op
```

`goos: linux
goarch: amd64
pkg: github.com/andrepputra/benchmarking
BenchmarkSimulateLootRNG-12         4659            261621 ns/op           18809 B/op        276 allocs/op
PASS
ok      github.com/andrepputra/benchmarking     2.308s`